requirements.txt
Pillow==2.6.1
PyBluez==0.18
argparse==1.2.1
cffi==0.8.6
chardet==2.3.0
colorama==0.3.2
cryptography==0.6.1
dnspython==1.12.0
evdev==0.4.7
gyp==0.1
html5lib==0.999
ndg-httpsclient==0.3.2
openrobertalab==1.4.0
ply==3.4
pyOpenSSL==0.14
pyasn1==0.1.7
pycparser==2.10
pycrypto==2.6.1
pygobject==3.14.0
pypng==0.0.18
python-ev3dev==VERSION
pyudev==0.16.1
requests==2.4.3
six==1.8.0
smbus==1.1
urllib3==1.9.1
wheel==0.24.0
wsgiref==0.1.2
